backend/requirements.txt
django~=3.2.8
djangorestframework~=3.13.1
mysqlclient~=2.1.0
eyeD3
django-cors-headers~=3.11.0
pillow~=9.0.1